Refactor/move UserMerger utility into admin/tool
